+++ spec/wavesurfer.spec.js
@@ -161,16 +161,16 @@ describe('WaveSurfer/playback:', function() {
 
     /** @test {WaveSurfer#skipForward}  */
     it('skip forward', function() {
-        // skip 4 seconds forward
-        wavesurfer.skipForward(4);
+        // skip x seconds forward
+        let expectedTime = 4;
+        wavesurfer.skipForward(expectedTime);
         let time = wavesurfer.getCurrentTime();
-        let expectedTime = 3.9999999999999996;
-        expect(time).toEqual(expectedTime);
+        expect(time).toBeNear(expectedTime, 0.0001);
 
         // skip forward with params.skipLength (default: 2 seconds)
         wavesurfer.skipForward();
         time = wavesurfer.getCurrentTime();
-        expect(time).toEqual(expectedTime + 2);
+        expect(time).toBeNear(expectedTime + 2, 0.0001);
     });
 
     /** @test {WaveSurfer#getPlaybackRate}  */
@@ -433,5 +433,17 @@ describe('WaveSurfer/errors:', function() {
 
             wave[0].setWaveColor('#000000');
         }).not.toThrow();
+    });
+
+    /**
+     * @test {WaveSurfer#load}
+     */
+    it('throw when url parameter for load is missing', function() {
+        var wave = TestHelpers.createWaveform({
+            container: '#test'
+        });
+        expect(function() {
+            wave[0].load();
+        }).toThrow(new Error('url parameter is invalid'));
     });
 });