+++ CHANGES.md
@@ -5,6 +5,8 @@ wavesurfer.js changelog
 ------------------
 
 - New `barRadius` option to create waveforms with rounded bars (#953)
+- Throw error when the url parameter is not supplied to
+  `wavesurfer.load()` (#1773)
 
 3.1.0 (26.09.2019)
 ------------------